Load all migrations paths for Run pending migrations button
Prior to this change, the PendingMigrationError only detected pending
migrations located in the top-level 'db/migrate' directory. Likewise,
clicking the "Run pending migrations" button in the browser would only
run the migrations from the top-level 'db/migrate' directory.

In contrast, when running `bin/rails db:migrate` from the command line,
all pending migrations were detected and run if they were located in a
configured migration path.

As a user, it is incredibly confusing to have the CLI do one thing and
the UI button do another. This is especially true given that the browser
error page is often the first place a developer sees that additional
migrations need to be run.

This change simply takes the `load_config` logic that is already in the
`db:migrate` task and applies it to the `ActiveRecord::Migration` and
`ActiveRecord::PendingMigrationError` classes to provide feature parity
between the CLI task and the UI button.